==> notebooks/old_scripts/0_DownloadData.ipynb <==
import os
from dotenv import load_dotenv
import aicrowd
from aicrowd import dataset

#load_dotenv('../.env')
#API_KEY = os.getenv('aicrowd_api')
#aicrowd.auth.aicrowd_login(api_key=API_KEY)

import pandas as pd
listData = pd.DataFrame(aicrowd.dataset.list_dataset('spotify-sequential-skip-prediction-challenge'))
listData


listData['filename'] = listData['title']
listData.loc[0,'filename'] = listData.loc[0,'filename']+'.pdf'
listData.loc[2,'filename'] = listData.loc[2,'filename']+'.pdf'
listData.loc[5,'filename'] = listData.loc[5,'filename']+'.tar'
listData

from multiprocessing.pool import ThreadPool
import requests

# Do not include the "Training_set" file as it will be downloaded in split files
for nFile in range(len(listData)-1):
    r = requests.get(listData.loc[nFile,'external_url'], stream=True)
    if r.status_code == requests.codes.ok:
        if '.pdf' in listData.loc[nFile,'filename']:
            saveFile = '../references/'+listData.loc[nFile,'filename']
        else:
            saveFile = '../data/raw/'+listData.loc[nFile,'filename']  
        print(saveFile)
        with open(saveFile, 'wb') as f:
            for data in r:
                f.write(data)


with open('../data/raw/Training_Set_Split_Download.txt') as f:
    targetURLs = f.read().splitlines() 
f.close()
    
for url in targetURLs:
    r = requests.get(url, stream=True)
    if r.status_code == requests.codes.ok:
        file_name_start_pos = url.rfind("/") + 1
        file_name = url[file_name_start_pos:]
        saveFile = '../data/raw/'+file_name
        print(saveFile)
        with open(saveFile, 'wb') as f:
            for data in r:
                f.write(data)
