==> notebooks/old_scripts/1_mini_DataWrangling.ipynb <==


# **Will you skip this music track or not?**


The public part of the dataset consists of roughly 130 million listening sessions with associated user interactions on the Spotify service. 

The task is to predict whether individual tracks encountered in a listening session will be skipped by a particular user. In order to do this, complete information about the first half of a user’s listening session is provided, while the prediction is to be carried out on the second half. Participants have access to metadata, as well as acoustic descriptors, for all the tracks encountered in listening sessions.

https://www.aicrowd.com/challenges/spotify-sequential-skip-prediction-challenge

Brost, B., Mehrotra, R., & Jehan, T. (2019, May). The music streaming sessions dataset. In The World Wide Web Conference (pp. 2594-2600).



As the entire dataset is too big to experiment data manipulation, Spotify provided a mini dataset for this purpose.

In this script, we will do data wrangling to inspect the quality of the data, and do data engineering to generate features for machine learning modeling.


# Mount Google drive to Colab


#from google.colab import drive
#drive.mount('/content/drive')
#%cd /content/drive/MyDrive/Capstone_SpotifyStreaming/notebooks

import numpy as np
import pandas as pd


# install pandas_profiling for auto data inspection

#import pandas_profiling
from pandas_profiling import ProfileReport 

# load the track information (mini version)

tf_df = pd.read_csv('../data/raw/data/track_features/tf_mini.csv')
tf_df.head()

tf_df.info()

# count the number of NaN values
tf_df.isnull().sum().sum()

# load the listening session data (mini version)

log_df = pd.read_csv('../data/raw/data/training_set/log_mini.csv')
log_df.head(20)

log_df.info()

# count the number of NaN values
tf_df.isnull().sum().sum()

# Generate ProfileReport of the track data

tf_profile = ProfileReport(tf_df, html={'style' : {'full_width':True}})
tf_profile.to_file(output_file="../reports/profiling/tf_mini_profile.html") 

#Download the profile report: https://drive.google.com/file/d/1-AbpYDsq1X-AGGl-2lHP-2ApgecaNAvJ/view?usp=share_link

# Generate ProfileReport of the listening session data

log_profile = ProfileReport(log_df, html={'style' : {'full_width':True}})
log_profile.to_file(output_file="../reports/profiling/log_mini_profile.html") 

#Download the profile report: https://drive.google.com/file/d/1-BRp061aKDdVDinAGXTUrdtb2UU-Z2dB/view?usp=share_link

According to the inspections above and the profile reports, there is no missing data, and the data type of each column was reasonablly coded. Therefore, there is no need to clean the data.
