==> notebooks/old_scripts/4-2_mini_FeatureEngineering.ipynb <==
# Check the results of Featuretools

import numpy as np
import pandas as pd
import featuretools as ft
import time
%matplotlib inline

# load the feature matrix to check whether it was properly saved.

feature_matrix2 = pd.read_csv('../data/processed/feature_matrix_skip2_TF.csv')
feature_matrix2.sort_values(by = ['session_id','session_position'],inplace=True)
feature_matrix2.head().T

#pd.set_option('display.max_rows', 1000)
pd.DataFrame(feature_matrix2.isna().sum())
#pd.set_option('display.max_rows', 60)

# inspect the data of a session
a = feature_matrix2[feature_matrix2['session_id'] == '0_00006f66-33e5-4de7-a324-2d18e439fc1e']
# inspect whether the featuretools generated features considered the session and the order information. The answer appears yes.
a[['tf.speechiness','CUM_SUM(tf.speechiness) by session_id','session_id','CUM_SUM(tf.speechiness) by session_id_skip_2_False','session_id_skip_2_False','CUM_SUM(tf.speechiness) by session_id_skip_2_True','session_id_skip_2_True','session_position']]

cols_skip_2_True = [col for col in feature_matrix2.columns if 'by session_id_skip_2_True' in col]
cols_skip_2_False = [col for col in feature_matrix2.columns if 'by session_id_skip_2_False' in col]


feature_matrix2.loc[feature_matrix2['session_id_skip_2_True'].isna(),cols_skip_2_True] = np.nan
feature_matrix2.loc[feature_matrix2['session_id_skip_2_False'].isna(),cols_skip_2_False] = np.nan
b = feature_matrix2[feature_matrix2['session_id'] == '0_00006f66-33e5-4de7-a324-2d18e439fc1e']
b[['tf.speechiness','CUM_SUM(tf.speechiness) by session_id','session_id','CUM_SUM(tf.speechiness) by session_id_skip_2_False','session_id_skip_2_False','CUM_SUM(tf.speechiness) by session_id_skip_2_True','session_id_skip_2_True','session_position']]

# shift the rows down by 1 within each session because it reflects the history (no_skip) up to the previous row
# without shifting, the information of no_skip will leak into the features!!
start_time = time.time()
feature_matrix2[cols_skip_2_True+cols_skip_2_False] = feature_matrix2.groupby(['session_id'])[cols_skip_2_True+cols_skip_2_False].shift(periods=1, axis=0, fill_value=0)
print('***It takes ',(time.time() - start_time)/60, ' mins.***')

start_time = time.time()

feature_matrix2 = feature_matrix2.groupby(['session_id'], sort = False).apply(lambda x: x.ffill())

#feature_matrix2.groupby(['session_id']).fillna(method='ffill', axis = 0, inplace=True)
print('***It takes ',(time.time() - start_time)/60, ' mins.***')

# check the data again after shifting the rows
b = feature_matrix2[feature_matrix2['session_id'] == '0_00006f66-33e5-4de7-a324-2d18e439fc1e']
# inspect whether the featuretools generated features considered the session and the order information. The answer appears yes.
b[['tf.speechiness','CUM_SUM(tf.speechiness) by session_id','session_id','CUM_SUM(tf.speechiness) by session_id_skip_2_False','session_id_skip_2_False','CUM_SUM(tf.speechiness) by session_id_skip_2_True','session_id_skip_2_True','session_position']]

feature_matrix2.to_csv('../data/processed/feature_matrix_skip2_TF_processed.csv')

# Scale standardization & training/testing data split will be performed at the model fitting stage.
