==> notebooks/old_scripts/7_data_splitting.ipynb <==
import glob
import random
import pandas as pd
n_file_valid = 2
valid_list = []
train_list = []
for n in range(10):
    file_list = glob.glob('../data/raw/training_set/log_'+str(n)+'*.csv')
    random.Random(23).shuffle(file_list)
    valid_list = valid_list + file_list[:n_file_valid]
    train_list = train_list + file_list[n_file_valid:]


pd.DataFrame(valid_list).to_csv('../data/interim/train_valid_split_1217/valid_file_list.csv', header = False, index = False)

pd.DataFrame(train_list).to_csv('../data/interim/train_valid_split_1217/train_file_list.csv', header = False, index = False)
