==> notebooks/old_scripts/9_calculate_item_similarities_for20180918-Copy1.ipynb <==
import pandas as pd
import numpy as np


# 7 days
# file_list = []
# for dateCode in range(911,918):
#     for logN in range(10):
#         file_list.append('../models/SVD/all_tracks/k300_v_log_'+str(logN)+'_20180'+str(dateCode)+'_000000000000.csv')
        
# file_list

# 1 day
file_list = []
for logN in range(10):
    file_list.append('../models/SVD/all_tracks/k300_v_log_'+str(logN)+'_20180917_000000000000.csv')
        
file_list

def cal_sim_svd(filename):
    v = pd.read_csv(filename).drop(columns=['Unnamed: 0'])
    v.columns = list(map(str, range(0,len(v)))) # make the column names 0 to 99 to match other data
    TrackTrackCosSim = v.transpose().dot(v) 
    TrackTrackLinCorr = v.corr()
    TrackTrackSpearCorr = v.corr(method = 'spearman')
    TrackTrackKendCorr = v.corr(method = 'kendall')
    
    from scipy.spatial.distance import cdist
    TrackTrackEuclDist = pd.DataFrame(cdist(v.T,v.T, 'euclidean'), index = TrackTrackCosSim.index, columns = TrackTrackCosSim.columns)
    TrackTrackManhDist = pd.DataFrame(cdist(v.T,v.T, 'cityblock'), index = TrackTrackCosSim.index, columns = TrackTrackCosSim.columns)
    TrackTrackSqEuclDist = pd.DataFrame(cdist(v.T,v.T, 'sqeuclidean'), index = TrackTrackCosSim.index, columns = TrackTrackCosSim.columns)
    TrackTrackCanbDist = pd.DataFrame(cdist(v.T,v.T, 'canberra'), index = TrackTrackCosSim.index, columns = TrackTrackCosSim.columns)
    TrackTrackHammDist = pd.DataFrame(cdist(v.T>0,v.T>0, 'hamming'), index = TrackTrackCosSim.index, columns = TrackTrackCosSim.columns)
    TrackTrackChebDist = pd.DataFrame(cdist(v.T,v.T, 'chebyshev'), index = TrackTrackCosSim.index, columns = TrackTrackCosSim.columns)
    TrackTrackBrayDist = pd.DataFrame(cdist(v.T,v.T, 'braycurtis'), index = TrackTrackCosSim.index, columns = TrackTrackCosSim.columns)

    
    
    sim_list = [TrackTrackCosSim, 
                TrackTrackLinCorr, 
                TrackTrackSpearCorr, 
                TrackTrackKendCorr, 
                TrackTrackEuclDist, 
                TrackTrackManhDist, 
                TrackTrackSqEuclDist, 
                TrackTrackCanbDist, 
                TrackTrackHammDist,
                TrackTrackChebDist,
                TrackTrackBrayDist]
    
    return sim_list
    

# from timeit import default_timer as timer #to see how long the computation will take
# import glob
# file_list = glob.glob('../models/SVD/all_tracks/k100_v*.csv')

# TrackTrackCosSim_mean = 0
# TrackTrackLinCorr_mean = 0
# TrackTrackSpearCorr_mean = 0
# TrackTrackKendCorr_mean = 0
# TrackTrackEuclDist_mean = 0
# TrackTrackManhDist_mean = 0
# TrackTrackSqEuclDist_mean = 0
# TrackTrackCanbDist_mean = 0
# TrackTrackHammDist_mean = 0
# TrackTrackChebDist_mean = 0
# TrackTrackBrayDist_mean = 0

# start = timer()
# for filename in file_list:
#     sim_list = []
#     sim_list = cal_sim_svd(filename)
#     TrackTrackCosSim_mean = sim_list[0]+TrackTrackCosSim_mean
#     TrackTrackLinCorr_mean = sim_list[1]+TrackTrackLinCorr_mean
#     TrackTrackSpearCorr_mean = sim_list[2]+TrackTrackSpearCorr_mean
#     TrackTrackKendCorr_mean = sim_list[3]+TrackTrackKendCorr_mean
#     TrackTrackEuclDist_mean = sim_list[4]+TrackTrackEuclDist_mean
#     TrackTrackManhDist_mean = sim_list[5]+TrackTrackManhDist_mean
#     TrackTrackSqEuclDist_mean = sim_list[6]+TrackTrackSqEuclDist_mean
#     TrackTrackCanbDist_mean = sim_list[7]+TrackTrackCanbDist_mean
#     TrackTrackHammDist_mean = sim_list[8]+TrackTrackHammDist_mean
#     TrackTrackChebDist_mean = sim_list[9]+TrackTrackChebDist_mean
#     TrackTrackBrayDist_mean = sim_list[10]+TrackTrackBrayDist_mean
    

# TrackTrackCosSim_mean = TrackTrackCosSim_mean/len(file_list)
# TrackTrackLinCorr_mean = TrackTrackLinCorr_mean/len(file_list)
# TrackTrackSpearCorr_mean = TrackTrackSpearCorr_mean/len(file_list)
# TrackTrackKendCorr_mean = TrackTrackKendCorr_mean/len(file_list)
# TrackTrackEuclDist_mean = TrackTrackEuclDist_mean/len(file_list)
# TrackTrackManhDist_mean = TrackTrackManhDist_mean/len(file_list)
# TrackTrackSqEuclDist_mean = TrackTrackSqEuclDist_mean/len(file_list)
# TrackTrackCanbDist_mean = TrackTrackCanbDist_mean/len(file_list)
# TrackTrackHammDist_mean = TrackTrackHammDist_mean/len(file_list)
# TrackTrackChebDist_mean = TrackTrackChebDist_mean/len(file_list)
# TrackTrackBrayDist_mean = TrackTrackBrayDist_mean/len(file_list)

# TrackTrackCosSim_mean.to_csv('../models/SVD/similarity/all_tracks/k100_CosSim.csv')
# TrackTrackLinCorr_mean.to_csv('../models/SVD/similarity/all_tracks/k100_LinCorr.csv')
# TrackTrackSpearCorr_mean.to_csv('../models/SVD/similarity/all_tracks/k100_SpearCorr.csv')
# TrackTrackKendCorr_mean.to_csv('../models/SVD/similarity/all_tracks/k100_KendCorr.csv')
# TrackTrackEuclDist_mean.to_csv('../models/SVD/similarity/all_tracks/k100_EuclDist.csv') # why all the same?
# TrackTrackManhDist_mean.to_csv('../models/SVD/similarity/all_tracks/k100_ManhDist.csv')
# TrackTrackSqEuclDist_mean.to_csv('../models/SVD/similarity/all_tracks/k100_SqEuclDist.csv') # why all the same?
# TrackTrackCanbDist_mean.to_csv('../models/SVD/similarity/all_tracks/k100_CanbDist.csv')
# TrackTrackHammDist_mean.to_csv('../models/SVD/similarity/all_tracks/k100_HammDist.csv')
# TrackTrackChebDist_mean.to_csv('../models/SVD/similarity/all_tracks/k100_ChebDist.csv')
# TrackTrackBrayDist_mean.to_csv('../models/SVD/similarity/all_tracks/k100_BrayDist.csv')

# print('Runtime: %0.2fs' % (timer() - start))

    

from timeit import default_timer as timer #to see how long the computation will take
# import glob
# file_list = glob.glob('../models/SVD/all_tracks/k300_v*.csv')

TrackTrackCosSim_mean = 0
TrackTrackLinCorr_mean = 0
TrackTrackSpearCorr_mean = 0
TrackTrackKendCorr_mean = 0
TrackTrackEuclDist_mean = 0
TrackTrackManhDist_mean = 0
TrackTrackSqEuclDist_mean = 0
TrackTrackCanbDist_mean = 0
TrackTrackHammDist_mean = 0
TrackTrackChebDist_mean = 0
TrackTrackBrayDist_mean = 0

start = timer()
for filename in file_list:
    sim_list = []
    sim_list = cal_sim_svd(filename)
    TrackTrackCosSim_mean = sim_list[0]+TrackTrackCosSim_mean
    TrackTrackLinCorr_mean = sim_list[1]+TrackTrackLinCorr_mean
    TrackTrackSpearCorr_mean = sim_list[2]+TrackTrackSpearCorr_mean
    TrackTrackKendCorr_mean = sim_list[3]+TrackTrackKendCorr_mean
    TrackTrackEuclDist_mean = sim_list[4]+TrackTrackEuclDist_mean
    TrackTrackManhDist_mean = sim_list[5]+TrackTrackManhDist_mean
    TrackTrackSqEuclDist_mean = sim_list[6]+TrackTrackSqEuclDist_mean
    TrackTrackCanbDist_mean = sim_list[7]+TrackTrackCanbDist_mean
    TrackTrackHammDist_mean = sim_list[8]+TrackTrackHammDist_mean
    TrackTrackChebDist_mean = sim_list[9]+TrackTrackChebDist_mean
    TrackTrackBrayDist_mean = sim_list[10]+TrackTrackBrayDist_mean
    

TrackTrackCosSim_mean = TrackTrackCosSim_mean/len(file_list)
TrackTrackLinCorr_mean = TrackTrackLinCorr_mean/len(file_list)
TrackTrackSpearCorr_mean = TrackTrackSpearCorr_mean/len(file_list)
TrackTrackKendCorr_mean = TrackTrackKendCorr_mean/len(file_list)
TrackTrackEuclDist_mean = TrackTrackEuclDist_mean/len(file_list)
TrackTrackManhDist_mean = TrackTrackManhDist_mean/len(file_list)
TrackTrackSqEuclDist_mean = TrackTrackSqEuclDist_mean/len(file_list)
TrackTrackCanbDist_mean = TrackTrackCanbDist_mean/len(file_list)
TrackTrackHammDist_mean = TrackTrackHammDist_mean/len(file_list)
TrackTrackChebDist_mean = TrackTrackChebDist_mean/len(file_list)
TrackTrackBrayDist_mean = TrackTrackBrayDist_mean/len(file_list)

# TrackTrackCosSim_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_CosSim.csv')
# TrackTrackLinCorr_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_LinCorr.csv')
# TrackTrackSpearCorr_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_SpearCorr.csv')
# TrackTrackKendCorr_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_KendCorr.csv')
# TrackTrackEuclDist_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_EuclDist.csv') # why all the same?
# TrackTrackManhDist_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_ManhDist.csv')
# TrackTrackSqEuclDist_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_SqEuclDist.csv') # why all the same?
# TrackTrackCanbDist_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_CanbDist.csv')
# TrackTrackHammDist_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_HammDist.csv')
# TrackTrackChebDist_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_ChebDist.csv')
# TrackTrackBrayDist_mean.to_csv('../models/SVD/all_tracks/similarity_for20180918/k300_BrayDist.csv')


TrackTrackCosSim_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_CosSim.csv')
TrackTrackLinCorr_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_LinCorr.csv')
TrackTrackSpearCorr_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_SpearCorr.csv')
TrackTrackKendCorr_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_KendCorr.csv')
TrackTrackEuclDist_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_EuclDist.csv') # why all the same?
TrackTrackManhDist_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_ManhDist.csv')
TrackTrackSqEuclDist_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_SqEuclDist.csv') # why all the same?
TrackTrackCanbDist_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_CanbDist.csv')
TrackTrackHammDist_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_HammDist.csv')
TrackTrackChebDist_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_ChebDist.csv')
TrackTrackBrayDist_mean.to_csv('../models/SVD/all_tracks/similarity_ave20180917/k300_BrayDist.csv')



# TrackTrackCosSim_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_CosSim.csv')
# TrackTrackLinCorr_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_LinCorr.csv')
# TrackTrackSpearCorr_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_SpearCorr.csv')
# TrackTrackKendCorr_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_KendCorr.csv')
# TrackTrackEuclDist_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_EuclDist.csv') # why all the same?
# TrackTrackManhDist_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_ManhDist.csv')
# TrackTrackSqEuclDist_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_SqEuclDist.csv') # why all the same?
# TrackTrackCanbDist_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_CanbDist.csv')
# TrackTrackHammDist_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_HammDist.csv')
# TrackTrackChebDist_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_ChebDist.csv')
# TrackTrackBrayDist_mean.to_csv('../models/SVD/all_tracks/similarity_train_valid_split/k300_BrayDist.csv')

print('Runtime: %0.2fs' % (timer() - start))

# # Fixing the index bug (Dec 6), no need to rerun it.
# import glob
# file_list = glob.glob('../models/SVD/similarity/*.csv')
# for file_name in file_list[0:1]:
#     df = pd.read_csv(file_name).drop(columns=['Unnamed: 0'])
#     df.columns = list(map(str, range(0,len(df))))
#     df.to_csv(file_name)


df.columns = list(map(str, range(0,len(df))))
df

a = np.array([[0, 0, 0],
              [0, 0, 1],
              [0, 1, 0],
              [0, 1, 1],
              [1, 0, 0],
              [1, 0, 1],
              [1, 1, 0],
              [1, 1, 1]])
pd.DataFrame(a)

from scipy.spatial.distance import cdist
pd.DataFrame(cdist(a.T,a.T, 'euclidean'))

TrackTrackCosSim = v.transpose().dot(v) 
TrackTrackCosSim.info()

TrackTrackCorr = v.corr()
TrackTrackCorr.info()
