==> notebooks/old_scripts/12-2_XGBoost_BayesianOpt.ipynb <==
import pandas as pd
import numpy as np
import tarfile
import io
import glob
import dask.dataframe as dd

import xgboost as xgb
from xgboost import XGBClassifier
from bayes_opt import BayesianOptimization

import warnings
warnings.simplefilter("ignore")



tar = tarfile.open('../data/raw/20181120_track_features.tar.gz', 'r:gz')
csv_files = tar.getnames()

df_list = []

for csv_file in [csv_files[2], csv_files[4]]:
    csv_contents = tar.extractfile(csv_file).read()
    df_list.append(pd.read_csv(io.BytesIO(csv_contents), encoding='utf8'))

tf_df = pd.concat(df_list, ignore_index=True)
tf_df.rename(columns={'track_id':'track_id_clean'}, inplace=True)

kmean100_df = pd.read_csv('../data/interim/all_data/mbKMeans100clusters.csv', usecols=['track_id','clus'])
kmean100_df.rename(columns={'track_id':'track_id_clean'}, inplace=True)



def get_sim(df_hist, df_lookup, sim_file_list, score_name_list):
    df_hist['ListenYes'] = (df_hist['skip_2'] == False)*1
    df_hist['ListenYes'].replace(0, -1, inplace = True)
    df_hist = df_hist.groupby(['session_id', 'clus']).agg({'ListenYes':['sum']})
    df_hist = df_hist.reset_index()
    df_hist.columns = df_hist.columns.droplevel(level = 1) # take out the unwanted level
    df_pivot = pd.pivot_table(df_hist, values = 'ListenYes',index='session_id', columns='clus')
    df_pivot = df_pivot.fillna(0)
    
    
    for sim_file, score_name in zip(sim_file_list, score_name_list):
        sim_matrix = pd.read_csv(sim_file).drop(columns=['Unnamed: 0'])
        sim_matrix.columns = list(map(str, range(0,len(sim_matrix))))
        df_sim_session = df_pivot.dot(sim_matrix)/sim_matrix.sum()
        
        df_lookup[score_name] = df_sim_session.lookup(df_lookup['session_id'],df_lookup['clus'].astype(str))
    
    return df_lookup

file_list = glob.glob('../data/raw/training_set/log_0*.csv')
file_list = file_list[0:round(len(file_list))]

df_lookup_list = []
for file in file_list:
    print(file)
    log_df = pd.read_csv(file)

    log_df = log_df[['session_id','track_id_clean','skip_2','session_position','session_length']].merge(kmean100_df)

    log_df_1 = log_df.loc[log_df['session_position']<(log_df['session_length']/2)]

    # as the entire dataset will be too big to train, 
    # train the first 2 tracks of the 2nd half (prediction set) should be enough, 
    # as these 2 contribute most to the spotify metric

    half_cut = log_df['session_length']/2

    log_df_2_1 = log_df.loc[(log_df['session_position']>=half_cut) & (log_df['session_position']<half_cut+1)]
    log_df_2_1['weight'] = 2
#     log_df_2_2 = log_df.loc[(log_df['session_position']>=half_cut+1) & (log_df['session_position']<half_cut+2)]
#     log_df_2_2['weight'] = 1

#     log_df_2 = pd.concat([log_df_2_1, log_df_2_2])
    log_df_2 = log_df_2_1


    sim_file_list = ['../models/SVD/similarity/k100_CanbDist.csv',
                     '../models/SVD/similarity/k100_CosSim.csv',
                     '../models/SVD/similarity/k100_LinCorr.csv',
                     '../models/SVD/similarity/k100_ManhDist.csv',
                     '../models/SVD/similarity/k100_HammDist.csv',
                     '../models/SVD/similarity/k100_SpearCorr.csv',
                     '../models/SVD/similarity/k100_KendCorr.csv']
    score_name_list = ['CanbDist100', 'CosSim100','LinCorr100','ManhDist100','HammDist100','SpearCorr100','KendCorr100']

    df_lookup_list.append(get_sim(log_df_1, log_df_2, sim_file_list, score_name_list))
    

df_lookup = pd.concat(df_lookup_list)

df_lookup = df_lookup.merge(tf_df)
df_lookup = pd.get_dummies(df_lookup, columns=['key','time_signature','mode'])

dtrain = xgb.DMatrix(df_lookup.drop(columns = ['session_id','track_id_clean','skip_2']), 
                     label=df_lookup['skip_2'],
                     weight = df_lookup['weight'])

def bo_tune_xgb(max_depth, gamma ,learning_rate, subsample, eta):
    params = {'max_depth': int(max_depth),
              'gamma': gamma,
              'learning_rate':learning_rate,
              'subsample': subsample,
              'eta': eta,
              'eval_metric': 'error'}
    #Cross validating with the specified parameters in 5 folds and 70 iterations
    cv_result = xgb.cv(params, dtrain, num_boost_round=100, nfold=5)
    #Return the negative RMSE
    return 1-cv_result['test-error-mean'].iloc[-1]

#Invoking the Bayesian Optimizer with the specified parameters to tune
xgb_bo = BayesianOptimization(bo_tune_xgb, {'max_depth': (3, 10),
                                             'gamma': (0, 1),
                                             'learning_rate':(0,0.01),
                                             'subsample': (0.3, 0.7),
                                             'eta':(0.1,0.3)
                                            })

from timeit import default_timer as timer #to see how long the computation will take
start = timer()
xgb_bo.maximize(n_iter=5, init_points=8, acq='ei')
print('Runtime: %0.2fs' % (timer() - start))

xgb_bo.max['params']

params = {'max_depth': 8,
              'gamma': 0.5899964093512567,
              'learning_rate':0.009417981929522201,
              'subsample': 0.8,
              'eta': 0.1,
              'eval_metric': 'error'}
cv_result = xgb.cv(params, dtrain, num_boost_round=70, nfold=5)


cv_result
